multiple worlds, raycast rotated/moved worlds
Adjusts the raycast to compensate for rotated or translated worlds.
Fixed multiple world support(I think)

+++ Assets/Voxelmetric/Code/World/World.cs
@@ -21,7 +21,7 @@ public class World : MonoBehaviour {
     public TerrainGen terrainGen;
 
     [HideInInspector]
-    public int worldIndex;
+    public byte worldIndex;
 
     void Start()
     {
